docling-service/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart
pydantic==2.9.2
pymupdf==1.24.9
pdfminer.six==20231228

